requirements.txt
mysqlclient
pyyaml
redis
